Added versions to all syncable instance on the server.  ChangeDescriptions now include a from and to versions.
All tests pass.

The client is not yet version-aware.  Since currently the client is connected
to exactly one server, there's an argument that the client could remain
unaware of instance versions.  Eventually, though, it'll probably be handy to
have instance versions on the client too (e.g. to recover from dropped
connections, or to facilitate client connections to multiple simultaneous servers,
or to reuse the protocol for partition communications, ...) But I'll wait a
bit until the need for versions more clear.