Interim version.
Added unpickling for some collections, and completed a test for pickling/unpickling
SyncableSeq to the RamPartition.

Changed some changes to SyncableReferences (with the kind), to get more symmetry
with the retrieval functions which should return SyncableReferences in
case the retriever wants to intanstiate a hollow one someday.  Also
simplifies some code in the RamPartition.

Fixed recent bug where commitToPartitions() was reversing the order of changes.
scala 2.8 has a new library function groupBy that might work for this.  Not
ready to make the move yet to 2.8 though.

Added MultiBuffer, which is similar to the std. library MultiMap.

Split test fixture into SyncFixture.  ObserveUtil is on its way out.