Implemented SyncableMap for completeness.  Currently requires that keys be Serializable, seems likely they'll be strings.
Tweaked the model for finding which vars are syncable properties.  Only vars in a
class that directly extends Syncable will be considered Properties.  So subclassing
a non-syncable class should work, and subclassing a syncable class should also
be possible.

Reworked reflection a bit, giving up on making it general purpose -- it'll
just be for Syncable now.

Inclued RamPartitionTests in master tests.