allow an optional 'options' as a second parameter when constructing syncable instances, so that I could pass a partition field. This saves a couple of lines calling withPartition() in the tests at least..
Some work on a Changes object, which is an event stream of changes to a property.
Next to each property is a Changes object, and another called $allChanges
that collects modification to all properties in that object.  For efficiency, though,
the Changes objects are created on demand.  The Changes for a property e.g. 'name'
are retrievable through .nameChanges().