save DataChanges as part of the transaction in Partition, then notify of changes if the commit succeeds.
Added RequestId to PickledWatch (in addition to ClientId).  The thought
is that ClientId will tend to be per machine or per connection, and so
persisting an extra identifier for the watches from one client
will be handy.
* used RequestId to identify individual functions for RamWatches

Implemented value type equality for SyncableId. (and SyncableReference
by inheritance)  This was handy for a test, but seems like the right
thing.
* (Can't make SyncableId a case class because its inherited, and also
  because we do more with the SyncableId object)

gave each SyncFixture a uniquely named test partition, which fixed
some broken tests.