Started implementing collection changes for SyncableSeq from partition Changes.
Fixed issue with Pickled, where watch operations were replacing PickledCollections
with Pickled instances

Added type parameter to PickledCollection in the hope of sharing some code.
it didn't help yet..

Added type parameters to TestRefObj so that library users wouldn't have to
cast as much.  Of course, then there's more type parameters to type so
it's kind of a wash.