Working on migration for pickled..
Added a new subclass of SyncableReference that has a kindVersion.  The
idea is to store this in the header of the Pickled.  That's next.

Reworked SyncableReference to make it a subclass of SyncableId, but not
also a container of a SyncableId

Started the convention the MigrateTo classes should implement copyTo
in reverse as well as forwards.  A good practice in case the migration
needs to be rolled back.