reworked commit() path for changes from the partition.  Cleaner now that we have queues to play with.  ProcessMessage should hopefully be able share the same commit path once client messages appear as ChangeDescriptions.
moved RamWatches trait into Partition in an attempt
to work around intermittent bytecode VerifyErrors in the build.

reworked the Id heirarchy to identify SyncNode: the client,
partition, or server app that could conceivably be sending
a series of changes.

tweak to the api for partition.watch: looks cleaner with a closure.