(interim version tests not passing) fixed generation of watches to the partition, but revealed issue where partition changes are being reflected back to origin.  I'll work on that next.
tweaked eclipse javascript formatter.

If we see a BeginWatch in ActiveSubscriptions, meaning that a new object has
been added to one of trees that the client is watching, we pickle the contents
of that object or collection immediately.  The logic for this is now
together in one place in ActiveSubscriptions.addedToSubscription.