(interim version)
Added SyncNode parameter for transactions, soon I'll try using that to stop
reflecting changes back to the submitter of the change.

Got rid of one of the two code paths for watch in partition.  Simplifies
the api which is nice.

Added some customObservePartition functions for testing partition observe
going through the normal app context commit mechanism.