(interim - all tests pass in first run, but some fail when run twice)
Modified core ProcessMessage.process to releases instance creation
and changes to the partition messages before observation changes.  It's
important that the creation changes come first, or the observe will
fail because it can't find the observed object.

Removed some Syncable initializers that set the value of properties.  For
now it's best to to not change anything, so that ChangeDescriptions are
not generated.  We can relax this rule later.

changed .js withTestSubscription to do a static version of watchDeep
when watching for server changes to test objects.  This required tweaking
one or two tests that were assuming two much about the change expected
to come back from the server.

load the iframe for the second browser tests on demand, instead of
statically when the page is loaded.

.js logging changes:
* added globalPrefix to the logging mechanism, used it to prefix the
  messages with the current Connection.  Handy for debugging the
* turned on printing of logger name by default for loggers created
  with $log.logger().  I may turn that back off soon though, it's
  useful when there's a lot of logging going on, but the added line
  noise is probably not worth it.

removed some no longer used AppContext.versioningDisabled cruft

thought about converting Message to a Seq[ChangeDescription] so that
I could use the cleaner code for applying a transaction, but backed off
for now.