Added test for modifying a seq in two browsers.
fixed issue when two second browser tests are run in sequence.  We were
creating one iframe for each and causing a collision.  Now we just
create one iframe for both tests.  The iframe is never destroyed, if
that starts to complicate debugging of subsequent tests, we can
figure out how to turn it off.

simplified some of the observation code in .js, by pssing in a function
to run while the observation is paused.  The server is moving away from
using a global pause on notification, perhaps the client should too.